fixing warnings refs #12
git-svn-id: file:///home/behr_mi/git/sumo_synched/trunk@12882 afbd958f-9f77-42d5-a016-97a22340ccf4

+++ sumo/src/guisim/GUIVehicle.cpp
@@ -1413,7 +1413,7 @@ GUIVehicle::getSeatPosition(size_t personIndex) const {
 int 
 GUIVehicle::getNumPassengers() const {
     if (myPersonDevice!=0) {
-        return myPersonDevice->getPersons().size();
+        return (int)myPersonDevice->getPersons().size();
     } 
     return 0;
 }
